windows-requirements.txt
-r requirements.txt
Pillow==6.2.0